Move `platform` module
