+++ desktop/detect/platform/index.js
