+++ desktop/menus/help-menu.js
@@ -1,7 +1,7 @@
 const { shell } = require('electron');
 
 const menuItems = require('./menu-items');
-const platform = require('../platform');
+const platform = require('../detect/platform');
 const build = require('../detect/build');
 
 const submenu = [
