+++ desktop/updater/lib/Updater.js
@@ -1,6 +1,6 @@
 const EventEmitter = require('events').EventEmitter;
 const { app, dialog, shell } = require('electron');
-const platform = require('../../platform');
+const platform = require('../../detect/platform');
 
 class Updater extends EventEmitter {
   constructor(changelogUrl, options) {
